requirements.txt
numpy>=1.14.0
torch==1.3.1
gym[atari]==0.12.5
tensorboardX==1.8
pytorch-transformers==1.1.0
tensorpack==0.9.4
matplotlib
opencv-python
gitpython==3.0.8
